requirements.txt
alembic==0.9.7
aniso8601==3.0.2
appnope==0.1.0
beautifulsoup4==4.7.1
boto3==1.8.1
botocore==1.11.1
bs4==0.0.1
cachetools==3.1.0
cairocffi==0.6
CairoSVG==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cssselect2==0.2.1
decorator==4.2.1
defusedxml==0.5.0
docutils==0.14
exotel==0.1.5
Flask==1.0.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
html5lib==1.0.1
httplib2==0.12.0
idna==2.6
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.4
mysqlclient==1.3.12
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.0
parso==0.1.1
pdfrw==0.4
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.18
Pygments==2.2.0
pygsheets==2.0.0
Pyphen==0.9.5
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
soupsieve==1.7.3
SQLAlchemy==1.2.2
tinycss2==0.6.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
